pyScript/requirements.txt
Flask==2.0.3
indic_transliteration==2.3.40
numpy==1.23.4
opencv_python==4.6.0.66
pdf2image==1.16.0
Pillow==9.3.0
pytesseract==0.3.10
selenium==4.6.0
TwoCaptcha==0.0.1
